requirements.txt
pygame
numpy
av
python-vlc